// frontend/src/generated/types/plantStatusResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface PlantStatusResponse {
  sysId?: string;
  plantCd?: string;
  plantNm?: string;
  paymentCd?: string;
  paymentNm?: string;
  lastCheckoutDate?: string;
  estimateCheckoutDate?: string;
}